gateway/requirements.txt
fastapi==0.68.1
uvicorn[standard]==0.15.0
python-multipart==0.0.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==4.0.2
prometheus-client==0.11.0
opentelemetry-api==1.7.1
opentelemetry-sdk==1.7.1
opentelemetry-instrumentation-fastapi==0.24b0
psutil==5.8.0
pillow==8.3.2
pydantic==1.8.2
python-dotenv==0.19.0
aiohttp==3.8.1 